# docs/openapi.v1.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Human Ticketing Console Control Plane API",
    "version": "1.0.0",
    "description": "Machine-readable contract for the local, human-in-the-loop REST control plane. Paths include the versioned protocol prefix and intentionally define no deployment address. The API exposes observation, lifecycle, reminders, explicit human confirmation, and redacted audit operations only."
  },
...
  "paths": {
...
    "/api/v1/schedules": {
      "get": {
        "operationId": "listSchedules",
        "tags": ["Schedules"],
        "summary": "List schedules ordered by start time",
        "parameters": [{ "$ref": "#/components/parameters/RequestIdHeader" }],
        "responses": {
          "200": { "$ref": "#/components/responses/ScheduleList" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "post": {
        "operationId": "createSchedule",
        "tags": ["Schedules"],
        "summary": "Create a local reminder schedule",
        "parameters": [
          { "$ref": "#/components/parameters/RequestIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" },
          { "$ref": "#/components/parameters/OperatorIdHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateScheduleRequest" } }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Schedule" },
          "403": { "$ref": "#/components/responses/PolicyDenied" },
          "409": { "$ref": "#/components/responses/ConflictError" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/schedules/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/ScheduleIdPath" }],
      "get": {
        "operationId": "getSchedule",
        "tags": ["Schedules"],
        "summary": "Read one schedule",
        "parameters": [{ "$ref": "#/components/parameters/RequestIdHeader" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Schedule" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "patch": {
        "operationId": "updateSchedule",
        "tags": ["Schedules"],
        "summary": "Update or cancel a schedule",
        "parameters": [
          { "$ref": "#/components/parameters/RequestIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" },
          { "$ref": "#/components/parameters/OperatorIdHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateScheduleRequest" } }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Schedule" },
          "403": { "$ref": "#/components/responses/PolicyDenied" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/ConflictError" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/schedules/{id}/acknowledge": {
      "parameters": [{ "$ref": "#/components/parameters/ScheduleIdPath" }],
      "post": {
        "operationId": "acknowledgeSchedule",
        "tags": ["Schedules"],
        "summary": "Record that a human saw a fired reminder",
        "parameters": [
          { "$ref": "#/components/parameters/RequestIdHeader" },
          { "$ref": "#/components/parameters/IdempotencyKeyHeader" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AcknowledgeScheduleRequest" } }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Schedule" },
          "403": { "$ref": "#/components/responses/PolicyDenied" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/ConflictError" },
          "422": { "$ref": "#/components/responses/ValidationError" },
          "503": { "$ref": "#/components/responses/ServiceUnavailable" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/v1/clock": {
      "get": {
        "operationId": "getClock",
        "tags": ["Schedules"],
        "summary": "Read the monotonic clock confidence snapshot",
        "parameters": [{ "$ref": "#/components/parameters/RequestIdHeader" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Clock" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "RequestIdHeader": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "description": "Optional caller correlation id. Invalid values are replaced with a generated id.",
        "schema": { "$ref": "#/components/schemas/RequestId" }
      },
      "IdempotencyKeyHeader": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": { "$ref": "#/components/schemas/IdempotencyKey" }
      },
      "OperatorIdHeader": {
        "name": "X-Operator-Id",
        "in": "header",
        "required": false,
        "description": "Audit identity for registration and schedule mutations; defaults to local-user.",
        "schema": { "$ref": "#/components/schemas/OperatorId" }
      },
...
      "ScheduleIdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "$ref": "#/components/schemas/UuidV4" }
      }
    },
    "headers": {
      "RequestId": {
        "description": "Correlation id echoed in the response body.",
        "schema": { "$ref": "#/components/schemas/RequestId" }
      }
    },
    "responses": {
...
      "Schedule": {
        "description": "Schedule snapshot.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleResponse" } } }
      },
      "ScheduleList": {
        "description": "Schedule collection.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleListResponse" } } }
      },
      "Clock": {
        "description": "Clock confidence snapshot.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ClockResponse" } } }
      },
...
      "PolicyDenied": {
        "description": "The operation or supplied content is outside policy.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
      },
      "NotFound": {
        "description": "The requested resource was not found.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
      },
      "ConflictError": {
        "description": "Current state, capacity, idempotency, or confirmation sequence conflicts with the request.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
      },
      "ValidationError": {
        "description": "Request path, query, header, or body validation failed.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
      },
...
      "ServiceUnavailable": {
        "description": "A bounded queue, cache, ticket store, or adapter is temporarily unavailable.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
      },
      "InternalError": {
        "description": "An unexpected internal error occurred without exposing internal details.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiErrorResponse" } } }
      }
    },
    "schemas": {
      "RequestId": {
        "type": "string",
        "minLength": 8,
        "maxLength": 128,
        "pattern": "^[A-Za-z0-9._:-]+$"
      },
      "IdempotencyKey": {
        "type": "string",
        "minLength": 8,
        "maxLength": 128,
        "pattern": "^[A-Za-z0-9._:-]+$"
      },
      "OperatorId": {
        "type": "string",
        "minLength": 1,
        "maxLength": 128,
        "pattern": "^[A-Za-z0-9._:-]+$"
      },
      "UuidV4": {
        "type": "string",
        "format": "uuid",
        "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
      },
...
      "DateTime": {
        "type": "string",
        "format": "date-time",
        "pattern": "(?:Z|[+-][0-9]{2}:[0-9]{2})$"
      },
...
      "Reminder": {
        "type": "object",
        "additionalProperties": false,
        "description": "Reminder targets must be strictly later than the control-plane current time when a schedule is created or updated.",
        "required": ["offset_seconds", "channel"],
        "properties": {
          "offset_seconds": { "type": "integer", "minimum": -604800, "maximum": 0 },
          "channel": { "type": "string", "enum": ["desktop", "sound"] }
        }
      },
      "ScheduleState": {
        "type": "string",
        "enum": ["draft", "scheduled", "notified", "human_confirmed", "completed", "failed", "cancelled", "expired"]
      },
      "ScheduleResource": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "label", "starts_at", "timezone", "reminders", "state", "created_at", "updated_at"],
        "properties": {
          "id": { "$ref": "#/components/schemas/UuidV4" },
          "label": { "type": "string", "minLength": 1, "maxLength": 128 },
          "public_reference": { "type": "string", "format": "uri", "pattern": "^https?://(?![^/?#]*@)[^?#]+$" },
          "starts_at": { "$ref": "#/components/schemas/DateTime" },
          "timezone": { "type": "string", "minLength": 1, "maxLength": 64 },
          "reminders": {
            "type": "array",
            "minItems": 1,
            "maxItems": 16,
            "items": { "$ref": "#/components/schemas/Reminder" }
          },
          "state": { "$ref": "#/components/schemas/ScheduleState" },
          "acknowledged_at": { "$ref": "#/components/schemas/DateTime" },
          "created_at": { "$ref": "#/components/schemas/DateTime" },
          "updated_at": { "$ref": "#/components/schemas/DateTime" }
        }
      },
      "CreateScheduleRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["label", "starts_at", "timezone", "reminders"],
        "properties": {
          "label": { "type": "string", "minLength": 1, "maxLength": 128 },
          "public_reference": { "type": "string", "format": "uri", "pattern": "^https?://(?![^/?#]*@)[^?#]+$" },
          "starts_at": { "$ref": "#/components/schemas/DateTime" },
          "timezone": { "type": "string", "minLength": 1, "maxLength": 64 },
          "reminders": {
            "type": "array",
            "minItems": 1,
            "maxItems": 16,
            "items": { "$ref": "#/components/schemas/Reminder" }
          }
        }
      },
      "UpdateScheduleRequest": {
        "description": "Use the cancelled branch alone for cancellation; label/reminders updates cannot be combined with state.",
        "oneOf": [
          {
            "type": "object",
            "additionalProperties": false,
            "required": ["state"],
            "properties": { "state": { "const": "cancelled" } }
          },
          {
            "type": "object",
            "additionalProperties": false,
            "minProperties": 1,
            "anyOf": [
              { "required": ["label"] },
              { "required": ["reminders"] }
            ],
            "properties": {
              "label": { "type": "string", "minLength": 1, "maxLength": 128 },
              "reminders": {
                "type": "array",
                "minItems": 1,
                "maxItems": 16,
                "items": { "$ref": "#/components/schemas/Reminder" }
              }
            }
          }
        ]
      },
      "AcknowledgeScheduleRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["operator_id"],
        "properties": { "operator_id": { "$ref": "#/components/schemas/OperatorId" } }
      },
...
      "ApiErrorCode": {
        "type": "string",
        "enum": [
          "policy.denied",
          "operator.confirmation_required",
          "command.expired",
          "command.stale",
          "idempotency.replay",
          "device.not_found",
          "device.busy",
          "adapter.unavailable",
          "clock.uncertain",
          "schema.invalid",
          "schedule.not_found",
          "schedule.not_notified",
          "schedule.started",
          "request.internal",
          "request.invalid"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["request_id", "code", "message"],
        "properties": {
          "request_id": { "$ref": "#/components/schemas/RequestId" },
          "code": { "$ref": "#/components/schemas/ApiErrorCode" },
          "message": { "type": "string", "minLength": 1 },
          "retryable": { "const": true },
          "details": { "type": "object", "additionalProperties": true }
        }
      },
...
      "ScheduleResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["request_id", "id", "label", "starts_at", "timezone", "reminders", "state", "created_at", "updated_at"],
        "properties": {
          "request_id": { "$ref": "#/components/schemas/RequestId" },
          "id": { "$ref": "#/components/schemas/UuidV4" },
          "label": { "type": "string", "minLength": 1, "maxLength": 128 },
          "public_reference": { "type": "string", "format": "uri", "pattern": "^https?://(?![^/?#]*@)[^?#]+$" },
          "starts_at": { "$ref": "#/components/schemas/DateTime" },
          "timezone": { "type": "string", "minLength": 1, "maxLength": 64 },
          "reminders": { "type": "array", "minItems": 1, "maxItems": 16, "items": { "$ref": "#/components/schemas/Reminder" } },
          "state": { "$ref": "#/components/schemas/ScheduleState" },
          "acknowledged_at": { "$ref": "#/components/schemas/DateTime" },
          "created_at": { "$ref": "#/components/schemas/DateTime" },
          "updated_at": { "$ref": "#/components/schemas/DateTime" }
        }
      },
      "ScheduleListResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["request_id", "items"],
        "properties": {
          "request_id": { "$ref": "#/components/schemas/RequestId" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/ScheduleResource" } }
        }
      },
      "ClockResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["request_id", "server_time", "monotonic_supported", "offset_ms", "confidence"],
        "properties": {
          "request_id": { "$ref": "#/components/schemas/RequestId" },
          "server_time": { "$ref": "#/components/schemas/DateTime" },
          "monotonic_supported": { "const": true },
          "offset_ms": { "type": "integer" },
          "confidence": { "type": "string", "enum": ["local", "uncertain"] }
        }
      },
...
    }
  }
}